+++ src/test/java/javax/util/streamex/CustomPoolTest.java
@@ -62,6 +62,7 @@ public class CustomPoolTest {
         assertArrayEquals(new Object[] {"a", "b", "c"}, StreamEx.of("a", "b", "c").parallel(pool).peek(this::checkThread).toArray());
         assertEquals("{ccc={bb={a={}}}}", StreamEx.of("a", "bb", "ccc").parallel(pool).peek(this::checkThread)
                 .foldLeft(Collections.emptyMap(), (acc, v) -> Collections.singletonMap(v, acc)).toString());
+        assertEquals(1000, IntStreamEx.constant(1, 1000).boxed().parallel(pool).peek(this::checkThread).foldLeft(0, Integer::sum).intValue());
     }
     
     @Test